requirements.txt
Django>=4.2
djangorestframework>=3.14
requests>=2.31
djangorestframework-simplejwt>=5.2
openai>=1.0
python-dotenv>=1.0 
google-generativeai>=0.3.0 